Refactor MicroservicesTable
